services/api/requirements.txt
# services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
strawberry-graphql[fastapi]==0.230.0
celery[redis]==5.4.0
redis[asyncio]==5.0.4
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-fastapi-instrumentator==6.1.0
boto3==1.34.0
python-multipart==0.0.9
